# crates/batchalign/batchalign-cli/openapi.json
{
  "components": {
    "schemas": {
      "BatchInferProgress": {
        "description": "Aggregate progress for a batched infer job across all language groups.",
        "properties": {
          "language_groups": {
            "additionalProperties": {
              "$ref": "#/components/schemas/LanguageGroupProgress"
            },
            "description": "Per-language-group progress, keyed by language code.\nUses BTreeMap for deterministic JSON serialization.",
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          }
        },
        "required": [
          "language_groups"
        ],
        "type": "object"
      },
...
      "ContentType": {
        "description": "Engine category that supports backend overrides.\n\nCurrently only ASR and FA have multiple engine backends.\nOther inference tasks (morphosyntax, utseg, translate, coref)\nalways use their single built-in engine.\nMIME-like content discriminator for file results.",
        "enum": [
          "chat",
          "csv",
          "text"
        ],
        "type": "string"
      },
...
      "DisplayPath": {
        "description": "Display path for a file within a job: either a bare basename\n(`\"sample.cha\"`) for single-file input or a relative forward-slash path\n(`\"PWA/TYO_a1.cha\"`) for directory input with subdirectories.\n\nBackslashes are normalized to forward slashes on construction so the value\nis platform-independent regardless of whether the CLI ran on Windows.\n\nThis type replaces the former `FileName` which incorrectly rejected path\nseparators during deserialization even though the system routinely carries\nrelative paths.",
        "type": "string"
      },
      "DurationSeconds": {
        "description": "Duration measured in fractional seconds.",
        "format": "double",
        "type": "number"
      },
      "ErrorResponse": {
        "description": "Standard error response body, matching FastAPI's `HTTPException` format.\n\nAll non-2xx responses from the server use this shape so that clients\ncan parse errors uniformly. The one exception is\n[`ServerError::JobConflict`](crate::error::ServerError::JobConflict),\nwhich nests a `conflicts` array inside `detail`.",
        "properties": {
          "detail": {
            "description": "Human-readable description of what went wrong. For validation\nerrors this includes the specific field or constraint that failed.",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "type": "object"
      },
...
      "ExecutionPlan": {
        "description": "Execution plan describing where and how a job is processed.\n\nAttached to `Job` and projected into `JobInfo` for API consumers.\nThe `stage` field tracks progress through the staged-remote lifecycle;\nfor local and explicit-remote jobs it is always [`ExecutionStage::Done`]\nonce the job finishes.",
        "properties": {
          "execution_host": {
            "description": "Hostname of the machine executing the job (e.g. `\"server-01\"`).",
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/ExecutionMode",
            "description": "How the job is being executed."
          },
          "remote_job_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JobId",
                "description": "For staged-remote jobs: the job ID on the remote server.\n`None` until the remote job is submitted."
              }
            ]
          },
          "stage": {
            "$ref": "#/components/schemas/ExecutionStage",
            "description": "Current lifecycle stage of the execution plan."
          }
        },
        "required": [
          "mode",
          "execution_host",
          "stage"
        ],
        "type": "object"
      },
...
      "FileResult": {
        "description": "Result for a single processed file.",
        "properties": {
          "content": {
            "description": "Processed file content (CHAT text or CSV).  Empty string when\n`error` is `Some`.",
            "type": "string"
          },
          "content_type": {
            "$ref": "#/components/schemas/ContentType",
            "description": "MIME-like content discriminator: `\"chat\"` for CHAT files (default),\n`\"csv\"` for tabular output (e.g. opensmile features)."
          },
          "error": {
            "description": "Human-readable error message if this file failed processing.\n`None` for successfully processed files.",
            "type": [
              "string",
              "null"
            ]
          },
          "filename": {
            "$ref": "#/components/schemas/DisplayPath",
            "description": "Display path for this file: a bare basename (`\"sample.cha\"`) or a\nrelative forward-slash path (`\"PWA/TYO_a1.cha\"`) for directory input.\nBackslashes are normalized to forward slashes on construction."
          },
          "provenance": {
            "description": "Processing provenance extracted from the output CHAT file.\nEach entry records one batchalign3 command that was applied\n(command name, engine version, timestamp). Empty for non-CHAT\noutput or files that failed processing.",
            "items": {
              "$ref": "#/components/schemas/ProvenanceEntry"
            },
            "type": "array"
          }
        },
        "required": [
          "filename"
        ],
        "type": "object"
      },
...
      "JobControlPlaneInfo": {
        "description": "Backend-owned orchestration metadata for a job projection.",
        "properties": {
          "backend": {
            "$ref": "#/components/schemas/JobControlPlaneBackendKind",
            "description": "Control-plane backend that produced this projection."
          },
          "temporal": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TemporalWorkflowExecutionInfo",
                "description": "Temporal workflow execution metadata when `backend == temporal`."
              }
            ]
          }
        },
        "required": [
          "backend"
        ],
        "type": "object"
      },
      "JobId": {
        "description": "Server-assigned identifier for a job (non-empty).",
        "type": "string"
      },
      "JobInfo": {
        "description": "`GET /jobs/{id}` response — job progress.\n\nContains full detail about a single job, including per-file statuses.\nPolled by the CLI to drive progress bars and by the dashboard for\nlive-updating tables.",
        "properties": {
          "active_lease": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LeaseRecord",
                "description": "Active lease information when this job is currently claimed by a node."
              }
            ]
          },
          "batch_progress": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BatchInferProgress",
                "description": "Per-language-group progress for batched text commands (morphotag,\nutseg, translate, coref).  Present only while a batch is in flight."
              }
            ]
          },
          "command": {
            "$ref": "#/components/schemas/ReleasedCommand",
            "description": "Batchalign command that was submitted (e.g. \"morphotag\", \"align\")."
          },
          "completed_at": {
            "description": "ISO 8601 timestamp of when the job reached a terminal state.",
            "type": [
              "string",
              "null"
            ]
          },
          "completed_files": {
            "description": "Number of files that have finished processing (status `Done`).\nInvariant: `0 <= completed_files <= total_files`.",
            "format": "int64",
            "type": "integer"
          },
          "control_plane": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JobControlPlaneInfo",
                "description": "Server control-plane metadata for this job when returned by a server."
              }
            ]
          },
          "current_file": {
            "description": "Filename currently being processed.  `None` when the job is queued\nor has reached a terminal state.",
            "type": [
              "string",
              "null"
            ]
          },
          "duration_s": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DurationSeconds",
                "description": "Wall-clock duration from first file dispatched to last file finished,\nin seconds.  `None` while the job is still active."
              }
            ]
          },
          "error": {
            "description": "Job-level error message (e.g. worker pool exhaustion, memory gate\ntimeout).  Distinct from per-file errors in `file_statuses`.",
            "type": [
              "string",
              "null"
            ]
          },
          "execution_plan": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ExecutionPlan",
                "description": "Execution plan describing where and how this job is processed.\nPresent for staged-remote jobs; `None` for local execution."
              }
            ]
          },
          "file_statuses": {
            "description": "Per-file processing status, one entry per submitted file.",
            "items": {
              "$ref": "#/components/schemas/FileStatusEntry"
            },
            "type": "array"
          },
          "job_id": {
            "$ref": "#/components/schemas/JobId",
            "description": "Server-assigned UUID (v4) for this job."
          },
          "lang": {
            "$ref": "#/components/schemas/LanguageSpec",
            "description": "Language specification: a resolved ISO 639-3 code or `\"auto\"`."
          },
          "last_cancelled_at": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UnixTimestamp",
                "description": "Wall-clock timestamp of the most recent cancel attempt against\nthis job. `None` until at least one cancel arrives. Denormalized\nfrom the `cancellations` audit table for fast list-view rendering\nwithout a JOIN. See `CancellationRecord` for the full audit\nhistory via `GET /jobs/{id}/cancellations`."
              }
            ]
          },
          "last_cancelled_host": {
            "description": "Caller-reported host of the most recent cancel attempt.",
            "type": [
              "string",
              "null"
            ]
          },
          "last_cancelled_reason": {
            "description": "Caller-reported reason of the most recent cancel attempt.",
            "type": [
              "string",
              "null"
            ]
          },
          "last_cancelled_source": {
            "description": "Wire-format source string of the most recent cancel attempt\n(`tui`, `api`, `signal`, ...). See `CancelSource`.",
            "type": [
              "string",
              "null"
            ]
          },
          "next_eligible_at": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UnixTimestamp",
                "description": "When the job is deferred before execution, the earliest unix timestamp\nat which it should be retried. `None` when the job is immediately\neligible to run."
              }
            ]
          },
          "num_workers": {
            "description": "Number of concurrent Python workers used for this job.  Determined\nby the pool at dispatch time.  `None` for queued jobs.",
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "options": {
            "description": "Typed submitted command options captured at job creation time.\n\nThe dashboard uses this to show the original argument/config payload for\ndebugging and operator review without reconstructing a lossy CLI string."
          },
          "source_dir": {
            "description": "Client's original input directory path, used for display in the\ndashboard and CLI output.  May be empty for content-mode submissions.",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "description": "Current lifecycle state of the job."
          },
          "submitted_at": {
            "description": "ISO 8601 timestamp of when the job was submitted (e.g.\n`\"2026-01-15T10:00:00Z\"`).",
            "type": [
              "string",
              "null"
            ]
          },
          "submitted_by": {
            "description": "IP address or Tailscale identifier of the submitting client.\nUsed for conflict detection (same `submitted_by` + filename =\nduplicate).",
            "type": [
              "string",
              "null"
            ]
          },
          "submitted_by_name": {
            "description": "Human-readable hostname of the submitting machine, resolved from\nTailscale or reverse DNS.  For dashboard display only.",
            "type": [
              "string",
              "null"
            ]
          },
          "total_files": {
            "description": "Total number of files in this job (immutable after submission).",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "job_id",
          "status",
          "command",
          "options",
          "total_files",
          "completed_files"
        ],
        "type": "object"
      },
...
      "JobResultResponse": {
        "description": "`GET /jobs/{id}/results` response — completed job results.",
        "properties": {
          "files": {
            "description": "Per-file results (empty until the job completes).",
            "items": {
              "$ref": "#/components/schemas/FileResult"
            },
            "type": "array"
          },
          "job_id": {
            "$ref": "#/components/schemas/JobId",
            "description": "Unique job identifier."
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "description": "Terminal status of the job."
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "type": "object"
      },
      "JobStatus": {
        "description": "Lifecycle states for a job — mirrors `JobStatus` enum.\n\nJobs progress through a linear lifecycle: `Queued -> Running -> {terminal}`.\nTerminal states (`Completed`, `Failed`, `Cancelled`, `Interrupted`) are\npermanent and never transition further.  Only `Cancelled` and `Failed` jobs\ncan be restarted (which resets them to `Queued`).",
        "enum": [
          "queued",
          "running",
          "completed",
          "failed",
          "cancelled",
          "interrupted",
          "writeback_failed"
        ],
        "type": "string"
      },
...
      "LanguageSpec": {
        "description": "Language specification from the CLI or job submission.\n\n`Auto` means the ASR engine should detect the language. This variant must\nbe resolved to a concrete [`LanguageCode3`] before any CHAT construction\nor NLP dispatch that requires a known language.\n\n`PerFile` means the command has no job-level language at all: each input\nfile's processing language is read from its `@Languages:` header at the\nstart of the per-file pipeline. This is distinct from `Auto`: `Auto` is an\nASR-engine signal asking the model to detect the spoken language;\n`PerFile` is a routing signal for text-NLP commands (morphotag, translate,\ncoref) whose language source is the CHAT file itself, not the job\nsubmission. The 2026-05-03 morphotag incident happened because these\ncommands were forced to carry a placeholder `Resolved(eng)` value that\nthen leaked into the job record, the dashboard, and the Stanza\npre-warming key. `PerFile` makes the absence of a job-level language a\nfirst-class state in the type system.",
        "oneOf": [
          {
            "description": "Let the ASR engine auto-detect the language.",
            "enum": [
              "Auto"
            ],
            "type": "string"
          },
          {
            "description": "A concrete ISO 639-3 language code.",
            "properties": {
              "Resolved": {
                "$ref": "#/components/schemas/LanguageCode3",
                "description": "A concrete ISO 639-3 language code."
              }
            },
            "required": [
              "Resolved"
            ],
            "type": "object"
          },
          {
            "description": "No job-level language; resolve per-file from each CHAT file's\n`@Languages:` header. Used by morphotag, translate, and coref —\nnone of which take a `--lang` CLI flag.",
            "enum": [
              "PerFile"
            ],
            "type": "string"
          }
        ]
      },
      "LeaseRecord": {
        "description": "Lease metadata for a claimed schedulable unit.\n\nIn single-node mode this is coarse-grained and attached at the job level:\nthe local dispatcher claims a queued job for a node, records the lease, and\nlater clears it when the runner exits. Fleet mode will extend the same shape\nwith real cross-node renewal and expiry handling.",
        "properties": {
          "expires_at": {
            "$ref": "#/components/schemas/UnixTimestamp",
            "description": "Unix timestamp when the lease should be considered expired if not renewed."
          },
          "heartbeat_at": {
            "$ref": "#/components/schemas/UnixTimestamp",
            "description": "Unix timestamp when the lease was created or most recently renewed."
          },
          "leased_by_node": {
            "$ref": "#/components/schemas/NodeId",
            "description": "Identifier of the node that currently owns the lease."
          }
        },
        "required": [
          "leased_by_node",
          "heartbeat_at",
          "expires_at"
        ],
        "type": "object"
      },
...
      "NodeId": {
        "description": "Identifier of a server/fleet node.\nEmpty when the node does not report an identity (older server versions).",
        "type": "string"
      },
...
      "ReleasedCommand": {
        "description": "Closed released command vocabulary used at all Rust seams.\n\nThis is the single canonical command type. Unknown command strings are\nrejected at deserialization boundaries (HTTP 422, DB recovery skip).",
        "enum": [
          "align",
          "transcribe",
          "transcribe_s",
          "translate",
          "morphotag",
          "coref",
          "utseg",
          "benchmark",
          "opensmile",
          "compare",
          "avqi"
        ],
        "type": "string"
      },
...
      "StatusMessageResponse": {
        "description": "Response body for mutating operations that return a status confirmation\n(e.g. cancel, delete, restart).\n\nReturned with HTTP 200 on success. Clients can branch on `status` for\nprogrammatic handling and display `message` to users.",
        "properties": {
          "message": {
            "description": "Human-readable sentence describing the action taken (e.g.\n`\"Job abc123 has been cancelled.\"`).",
            "type": "string"
          },
          "status": {
            "description": "Machine-readable outcome token (e.g. `\"cancelled\"`, `\"deleted\"`,\n`\"restarted\"`). Suitable for switch/match in client code.",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "type": "object"
      },
...
      "UnixTimestamp": {
        "description": "Unix timestamp as fractional seconds since epoch.",
        "format": "double",
        "type": "number"
      },
...
    }
  },
  "info": {
    "description": "batchalign3: command-line interface, application, server, worker pool, cache",
    "license": {
      "identifier": "BSD-3-Clause",
      "name": "BSD-3-Clause"
    },
    "title": "batchalign",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/jobs/{job_id}": {
      "delete": {
        "description": "Returns 409 if the job is still running -- the caller must cancel it first.\nDeleting a job removes it from the in-memory store and SQLite, and broadcasts\na `JobDeleted` event to connected WebSocket/SSE clients.",
        "operationId": "delete_job",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusMessageResponse"
                }
              }
            },
            "description": "Job deleted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job still running"
          }
        },
        "summary": "Permanently remove a terminal job and its associated state.",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "description": "This is the primary polling endpoint for the CLI: the client calls it\nrepeatedly until `status` reaches a terminal value.",
        "operationId": "get_job",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobInfo"
                }
              }
            },
            "description": "Job detail"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          }
        },
        "summary": "Return full detail for a single job, including per-file statuses.",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/jobs/{job_id}/results": {
      "get": {
        "description": "Blocks until the job has completed, failed, or been interrupted -- returns\n409 for jobs that are still running or queued. The response includes the\nfull staged content of each successful output file. Paths-mode jobs keep\nthe execution-host write, but also preserve a staged copy for download.\nCancelled jobs are intentionally excluded because they may have incomplete\nresults.",
        "operationId": "get_results",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResultResponse"
                }
              }
            },
            "description": "Results for terminal jobs"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          }
        },
        "summary": "Retrieve all output files for a terminal job.",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/results/{*filename}": {
      "get": {
        "description": "Waits for the individual file to reach a terminal status, not the whole job,\nso clients can fetch results incrementally as files finish. Supports\nstem-based matching (e.g., requesting `sample.wav` will find `sample.cha`)\nto handle the common case where the output extension differs from the input.",
        "operationId": "get_single_result",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Requested output filename",
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileResult"
                }
              }
            },
            "description": "One file result"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "File not found"
          }
        },
        "summary": "Retrieve the output for a single file within a job.",
        "tags": [
          "jobs"
        ]
      }
    },
...
  },
...
}